fix exit issue, call rootTunnel to exit
